requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
celery==5.6.2
celery-types==0.23.0
click==8.1.8
colorlog==6.10.1
DateTime==5.5
execnet==2.1.2
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.3
importlib_metadata==8.7.1
iniconfig==2.1.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
packaging==26.0
pipreqs==0.5.0
platformdirs==4.4.0
playwright==1.58.0
pluggy==1.6.0
pure_eval==0.2.3
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
pytest==8.4.2
pytest-base-url==2.1.0
pytest-check==2.6.2
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-playwright==0.7.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
referencing==0.36.2
requests==2.32.5
selenium==4.36.0
text-unidecode==1.3
tinycss2==1.4.0
urllib3==1.26.20

